+++ src/queue.rs
@@ -80,6 +80,7 @@ pub enum AppTabs {
 	Files,
 	Stashing,
 	Stashlist,
+	Graph,
 }
 
 ///